requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
pymongo==4.6.1
asyncio==3.4.3
logging==0.4.9.6 